split dynatrace topology and event check
