+++ dynatrace_event/CHANGELOG.md
@@ -1,0 +1,2 @@
+# CHANGELOG - Dynatrace_event
+
