+++ dynatrace_event/MANIFEST.in
@@ -1,0 +1,10 @@
+graft stackstate_checks
+graft tests
+
+include MANIFEST.in
+include README.md
+include requirements.in
+include requirements-dev.txt
+include manifest.json
+
+global-exclude *.py[cod] __pycache__
