+++ dynatrace_event/manifest.json
@@ -1,0 +1,23 @@
+{
+  "display_name": "Dynatrace Event",
+  "maintainer": "info@stackstate.com",
+  "manifest_version": "1.0.0",
+  "name": "dynatrace_event",
+  "metric_prefix": "dynatrace_event.",
+  "metric_to_check": "",
+  "creates_events": false,
+  "short_description": "Collect events from Dynatrace",
+  "guid": "9808ba89-2869-4078-bdd9-912efbd8ec2e",
+  "support": "core",
+  "supported_os": [
+    "linux",
+    "mac_os",
+    "windows"
+  ],
+  "public_title": "StackState-Dynatrace_event Integration",
+  "categories": [
+    ""
+  ],
+  "type": "check",
+  "is_public": false
+}
