+++ dynatrace_event/requirements-dev.txt
@@ -1,0 +1,1 @@
+-e ../stackstate_checks_dev
