+++ dynatrace_event/stackstate_checks/__init__.py
@@ -1,0 +1,4 @@
+# (C) StackState 2020
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+__path__ = __import__('pkgutil').extend_path(__path__, __name__)
