+++ dynatrace_event/stackstate_checks/dynatrace_event/__about__.py
@@ -1,0 +1,4 @@
+# (C) StackState 2020
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+__version__ = '0.0.1'
