+++ dynatrace_event/stackstate_checks/dynatrace_event/__init__.py
@@ -1,0 +1,10 @@
+# (C) StackState 2020
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+from .__about__ import __version__
+from .dynatrace_event import DynatraceEventCheck
+
+__all__ = [
+    '__version__',
+    'DynatraceEventCheck'
+]
