+++ dynatrace_event/tests/__init__.py
@@ -1,0 +1,3 @@
+# (C) StackState 2020
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
