+++ dynatrace_event/tox.ini
@@ -1,0 +1,29 @@
+[tox]
+minversion = 2.0
+skip_missing_interpreters = true
+basepython = py27
+envlist =
+    {py27,py36}-dynatrace_event
+    flake8
+
+[testenv]
+usedevelop = true
+platform = linux|darwin|win32
+deps =
+    -e../stackstate_checks_base[deps]
+    -rrequirements-dev.txt
+passenv =
+    DOCKER*
+    COMPOSE*
+commands =
+    pip install -r requirements.in
+    pytest -v
+
+[testenv:flake8]
+skip_install = true
+deps = flake8
+commands = flake8 .
+
+[flake8]
+exclude = .eggs,.tox,build
+max-line-length = 120
